Adding Constants File to clean up code
