+++ main/Constants.py
@@ -1,0 +1,15 @@
+# Purpose:
+# This file is for storing or updating constant values that may need to be changes depending on system requirements and
+# different usecases.
+removeList = [".gitignore", ".DS_Store", "desktop.ini"]
+
+class bcolors:
+    """
+    The colour of console messages.
+    """
+
+    OK = '\033[92m'  # GREEN
+    WARNING = '\033[93m'  # YELLOW
+    ERROR = '\033[91m'  # RED
+    RESET = '\033[0m'  # RESET COLOR
+
