+++ main/DNA_Generator.py
@@ -11,16 +11,7 @@ import random
 from functools import partial
 from .loading_animation import Loader
 from . import Rarity, Logic, Checks
-
-
-class bcolors:
-   """
-   The colour of console messages.
-   """
-   OK = '\033[92m'  # GREEN
-   WARNING = '\033[93m'  # YELLOW
-   ERROR = '\033[91m'  # RED
-   RESET = '\033[0m'  # RESET COLOR
+from .Constants import bcolors, removeList
 
 
 def get_hierarchy():
