+++ main/get_combinations.py
@@ -2,15 +2,6 @@ import bpy
 
 from . import DNA_Generator
 
-class bcolors:
-   """
-   The colour of console messages.
-   """
-   OK = '\033[92m'  # GREEN
-   WARNING = '\033[93m'  # YELLOW
-   ERROR = '\033[91m'  # RED
-   RESET = '\033[0m'  # RESET COLOR
-
 
 def get_combinations():
    """
